requirements.txt
fastapi
uvicorn
requests
PyMuPDF
Pillow
python-multipart
python-dotenv